Update postman collection and move classes
