+++ UserService.UnitTests/AddressServiceTests.cs
@@ -10,8 +10,8 @@ using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using AddressService.Core.Contracts;
 using UserService.Core.Config;
+using UserService.Core.Domains.Entities;
 using UserService.Core.Interfaces.Utils;
 using UserService.Core.Services;
