+++ UserService.UnitTests/GetHelperCoordsByPostcodeAndRadiusHandlerTests.cs
@@ -1,5 +1,4 @@
-﻿using AddressService.Core.Contracts;
-using Moq;
+﻿using Moq;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
@@ -7,7 +6,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using UserService.Core;
-using UserService.Core.Contracts;
+using UserService.Core.Domains.Entities;
 using UserService.Core.Dto;
 using UserService.Core.Interfaces.Services;
 using UserService.Core.Utils;
