+++ UserService/UserService.Core/Services/AddressService.cs
@@ -8,8 +8,8 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using AddressService.Core.Contracts;
 using UserService.Core.Config;
+using UserService.Core.Domains.Entities;
 using UserService.Core.Interfaces.Services;
 using UserService.Core.Interfaces.Utils;
 using UserService.Core.Utils;
