+++ UserService/UserService.Core/UserService.Core.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="Microsoft.Extensions.Http" Version="2.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Contracts\" />
+  </ItemGroup>
+
 </Project>
