+++ UserService/UserService.Handlers/GetHelperCoordsByPostcodeAndRadiusHandler.cs
@@ -1,12 +1,11 @@
-﻿using AddressService.Core.Contracts;
-using HelpMyStreet.Utils.Utils;
+﻿using HelpMyStreet.Utils.Utils;
 using MediatR;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using UserService.Core;
-using UserService.Core.Contracts;
+using UserService.Core.Domains.Entities;
 using UserService.Core.Dto;
 using UserService.Core.Interfaces.Services;
 using UserService.Core.Utils;
